Async `with_base_dir` implementation in `LocalFS` (#10)
This PR modifies `LocalFS::with_base_dir` to use an async version of the
`is_dir` check instead of the sync version it previously used. It also
adds a `LocalFS::with_base_dir_unchecked` function that does not do an
`is_dir` check.

Fixes #9

### Breaking change rationale

This is technically a breaking change (as it makes a sync function into
an async one), but I'm okay merging it for the following reasons:

- The project does not have many dependents yet (and the external ones
I've found do not use `with_base_dir`).
- The break is fairly easy to fix (either switch to `with_base_dir` if
you can easily use an async function or manually call `.is_dir()` on the
path before you pass it into `with_base_dir_unchecked`)

For a project at this stage, this choice seems to make sense compared to
the overhead of deprecating the function, introducing an async
alternative, and waiting for a major release to remove the sync one
(especially since any breakages will be found when doing a `cargo
update` instead of during a random build).