Ensure `read_link` returns "absolute" paths (#12)
Lunchbox uses the `relative_path` crate as a platform independent path
library (instead of implementing one inside Lunchbox). As you might
expect, `relative_path` doesn't support absolute paths.

This is a problem for symlinks.

On Unix systems, `read_link` generally returns the symlink target path
that was stored. This could be relative to the symlink file or an
absolute path.

Because the `relative_path` library doesn't give us a way to distinguish
between those, we must enforce that `read_link` always returns a path
relative to the filesystem root.

Hopefully we can improve this in the future.